requirements.txt
# Runtime dependencies
Flask==1.1.1
Flask-API==1.1
Flask-SQLAlchemy==2.4.1
PyMySQL==0.9.3
psycopg2-binary==2.8.3

# Testing
nose==1.3.7
pinocchio==0.4.2
factory-boy==2.12.0
coverage==4.5.4
pylint>=2.4.1

# Sentiment analysis
tweepy==3.8.0 
nltk==3.4.5 
google-cloud-language==1.3.0